+++ stubs/SilverStripe/ORM/DataObject.php
@@ -1,10 +1,12 @@
 <?php
 
 namespace SilverStripe\ORM;
+use SilverStripe\Core\Config\Configurable;
+
 if (class_exists('SilverStripe\ORM\DataObject')) {
     return;
 }
 class DataObject
 {
-
+    use Configurable;
 }